refactor: add helper function to help display old/new values
