+++ frontend/src/components/ChangeRequests/StatusChangeReviewCard/StatusChangeReviewCard.jsx
@@ -1,4 +1,5 @@
 import PropTypes from "prop-types";
+import { renderChangeValues } from "../../../helpers/changeRequests.helpers";
 import { renderField } from "../../../helpers/utils";
 import { useGetBLITotal, useGetNameForCanId } from "../../../hooks/lookup.hooks";
 import ReviewCard from "../ReviewCard";
@@ -16,39 +17,12 @@ import TermTag from "../TermTag";
  * @returns {JSX.Element} - The rendered component
  */
 function StatusChangeReviewCard({ agreementId, requesterName, requestDate, bliId, changeTo }) {
-    const KEY_NAMES = {
-        AMOUNT: "amount",
-        CAN: "can_id",
-        DATE_NEEDED: "date_needed",
-        STATUS: "status"
-    };
     const keyName = Object.keys(changeTo)[0];
-    let oldValue,
-        newValue = "";
     const totalAmount = useGetBLITotal(bliId);
     const oldCan = useGetNameForCanId(changeTo.can_id?.old);
     const newCan = useGetNameForCanId(changeTo.can_id?.new);
+    const { oldValue, newValue } = renderChangeValues(keyName, changeTo, oldCan, newCan);
 
-    switch (keyName) {
-        case KEY_NAMES.AMOUNT:
-            oldValue = renderField(keyName, "amount", changeTo.amount.old);
-            newValue = renderField(keyName, "amount", changeTo.amount.new);
-            break;
-        case KEY_NAMES.CAN:
-            oldValue = oldCan;
-            newValue = newCan;
-            break;
-        case KEY_NAMES.DATE_NEEDED:
-            oldValue = renderField(keyName, "date_needed", changeTo.date_needed.old);
-            newValue = renderField(keyName, "date_needed", changeTo.date_needed.new);
-            break;
-        case KEY_NAMES.STATUS:
-            oldValue = renderField(keyName, "status", changeTo.status.old);
-            newValue = renderField(keyName, "status", changeTo.status.new);
-            break;
-        default:
-            break;
-    }
     return (
         <ReviewCard
             type="Status Change"
